fix clerk auth to use the new auth pattern
